fix(web): 恢复详情侧栏导航的滚动与对齐
DetailNavSidebar 抽取时把整个组件(返回/身份/搜索/导航)放进了 layout 的
header 插槽 —— 该插槽是裸 shrink-0 容器,既不加 padding 也不滚动。导航由
此丢掉 content 插槽提供的两样东西:ScrollArea(tab 一多底部被裁、无法滚
动)和 p-2 叠加(hover inset 16px → 8px,与 px-4 的头部块失对齐)。

拆成两件恢复原结构:DetailNavSidebar 只保留固定的头部块(返回/身份/搜
索),新增 DetailNavMenu 渲染分组导航、放回带 ScrollArea 的 content 插槽;
搜索词通过 v-model 抬到调用方共享,选中即清空的行为不变。bot 与 project
两个详情页同步接线。
